Amend scoring and end round
This commit amends the scoring system so that each answer is worth one
extra point for each player to have got an answer wrong since the
current player's last answer, as opposed to just one extra point for
each consecutive wrong answer prior to this one.

Also included in this commit is functionality for ending the round when
either all correct answers have been given, or when every player has got
a wrong answer consecutively. In the latter case, the "Go" button
reveals the remaining answers one at a time.