Ensure consistent ordering in world cup svg script
This commit ensures that in each iteration of the animation script in
the world cup svg, the set of countries to be updated is sorted. This
ensures that the ordering of these statements will always be consistent,
and avoids the issue of multiple redundant diff lines each time a new
version is generated and commited.