Introduce concept of "break" level
This commit introduces the concept of a "break" level. When the
tournament goes into a break level, the clock automatically pauses and
the timer should be manually started when the last table has finished
play. This ensures that all players get at least the allotted amount of
break time.

This commit also changes the language on the screen from "next blinds"
to "next level", to reflect that a level can be a break and not just a
blinds configuration.