Implement seating assignment
This commit implements seating assignment as a new step between setting
up and starting the game. Instead of "Start game", the button on the
set-up screen is now "Assign seats", and, when clicked, we show the new
seating chart container. This will allow us to show the seating chart on
the screen before starting the game. On clicking "Start game" (now
displayed within the seating chart container), the clock will appear on
the screen and players can begin play.