Ensure timer is accurate
This commit changes the timing logic to ensure it is accurate.
Previously we were relying on `setTimeout` to ensure that the timer
ticked down by one second every 1000 milliseconds, but this was
extremely unreliable and would have ended up with rounds lasting far
longer than they should have done.

Now we set the `level_end` timestamp when we start a level or unpause
the timer, invoke the `tick` function multiple times each second, and
update the `minutes` and `seconds` variables based on the difference
between the current time and the `level_end` timestamp.