Remove phantom second
This commit fixes an issue where, when the level ticks over, we see the
full level duration for a split second before it ticks down to one
second less. This means that the clock ticks down twice in quick
succession and looks weird, right at the time people will be looking at
the screen. With this fix, the clock visually ticks down from 00:00 to
xx:59 while still functionally maintaining the correct full level
duration.