Fix bug with starting game
The previous commit, using a defined timestamp for level end, created a
bug where the `start_game` function failed because it was written to
start at level `-1` with the time at zero and then utilise the `tick`
function to tick over into level `0`. This is fixed by calling
`set_level_end` after we set `minutes` and `seconds` to zero.