Implement seating manager
This commit implements a tool to manage the seating during a tournament.
It is opened from the tournament clock using the `open_related` function
(previously `open_document`, now generalised), and is thus initialised
with the seating arrangement drawn by the tournament clock tool.

As this tool uses the `players_per_table` property to determine how many
seats to render in each table, the default value for this is changed
from `5` to `6`. It was never really intended to be `5`, but the idea
was to ensure there would be free seats on every table for late
registrations. This can be managed easily enough by simply starting a
new table if required.