Amend default blinds and prize money distribution
This commit amends the default blinds to go up more quickly, and the
prize money distribution to remain winner-takes-all up to six players.
This is based on the experience of the first Cameron's Bar event. Also
in this commit is a fix for an issue where a blank blinds level was
being created at the end of the levels due to there being an empty
string following the final semi-colon.