Minor tweaks to World Cup animation
This commit makes some minor tweaks to the World Cup animation.

Firstly the label is moved right into the bottom-left corner, where it
will look more natural than being in the middle of the screen.

Secondly the period of time during which the animation pauses after
completion before restarting, is increased from 2.5 seconds to 5
seconds.