Support simulation with pots already known
This commit adds support to the world cup draw simulation for entering
the order of pots already drawn. Also in this commit is the
implementation for two new extractors, for extracting the group and the
third pot team. Once two pots have been drawn, this will be useful to
get an up-to-date picture of how likely the UEFA pot 4 team is to be
drawn in the groups, at a point in time when we know some of the teams
in those groups.